pyproject.toml
[project]
name = "e2e-aichallenge-playground"
version = "0.1.0"
description = "End-to-end AI Challenge Playground"
requires-python = ">=3.12"
dependencies = [
    "experiment",
    "deptry>=0.12.0",
    "pre-commit>=3.0.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "lark>=1.0.0",
    "ruff>=0.14.10",
    "pure-pursuit",
    "mppi-planner",
    "mppi-controller",
    "planning-utils",
    "pid-controller",
    "tiny-lidar-net",
    "simulator",
    "core",
    "ad-component-core",
    "torch>=2.9.1",
    "mcap>=1.3.0",
    "mcap-ros2-support>=0.5.6",
    "tqdm>=4.67.1",
    "hydra-core>=1.3.2",
    "hydra-joblib-launcher>=1.2.0",
    "wandb>=0.23.1",
    "py-spy>=0.4.1",
    "ideal-sensor",
    "mlflow>=3.7.0",
    "dvc[s3]>=3.65.0",
    "numba>=0.63.1",
    "onnxscript>=0.5.7",
    "onnxscript>=0.5.7",
    "lateral-shift-planner",
    "pure-pursuit-controller",
    "matplotlib>=3.5.0",
    "scipy>=1.10.0",
    "plotly>=5.0.0",
    "rosbags>=0.9.15",
    "matplotlib>=3.5.0",
    "scipy>=1.10.0",
    "plotly>=5.0.0",
    "rosbags>=0.9.15",
    "cvxpy>=1.4.0",
    "mpc-lateral-controller",
]

[tool.uv]
managed = true

[tool.uv.sources]
experiment = { workspace = true }
core = { workspace = true }
ad-component-core = { workspace = true }
pure-pursuit = { workspace = true }
centerline-planner = { workspace = true }
pure-pursuit-controller = { workspace = true }
mppi-planner = { workspace = true }
mppi-controller = { workspace = true }
planning-utils = { workspace = true }
pid-controller = { workspace = true }
tiny-lidar-net = { workspace = true }
simulator = { workspace = true }
"ideal-sensor" = { workspace = true }
"dashboard" = { workspace = true }
"supervisor" = { workspace = true }
"logger" = { workspace = true }
"lateral-shift-planner" = { workspace = true }
"mpc-lateral-controller" = { workspace = true }

[tool.uv.workspace]
members = [
    "core",
    "ad_components/planning/lateral_shift_planner",
    "ad_components/core",
    "ad_components/planning/pure_pursuit",
    "ad_components/planning/centerline_planner",
    "ad_components/planning/planning_utils",
    "ad_components/planning/mppi_planner",
    "ad_components/control/mppi_controller",
    "ad_components/control/pid_controller",
    "ad_components/control/pure_pursuit_controller",
    "ad_components/control/tiny_lidar_net",
    "simulator",
    "experiment",
    "experiment/training",
    "dashboard",
    "supervisor",
    "logger",
    "ad_components/perception/ideal_sensor",
    "ad_components/control/mpc_lateral_controller",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SLF",    # flake8-self (private member access detection)
    "SIM",    # flake8-simplify
    "RUF",    # Ruff-specific rules
]

ignore = [
    "E501",   # Line too long (handled by formatter)
    "SIM108", # Use ternary operator (allow if-else for readability)
    "ERA001", # Commented-out code (false positives on Japanese comments)
    "RUF001", # Ambiguous unicode characters in strings
    "RUF002", # Ambiguous unicode characters in comments
    "RUF003", # Ambiguous unicode characters in docstrings
]

[tool.ruff.lint.per-file-ignores]
"**/tests/**" = ["SLF001"]  # Allow private access in tests
"core/src/core/data/**" = ["ERA001"]  # Allow Japanese section comments

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false # Start lenient
ignore_missing_imports = true

[tool.pytest.ini_options]
# Disable ROS pytest plugins that interfere with standard pytest
addopts = "-p no:launch_testing -p no:launch_testing_ros -p no:launch_testing_ros_pytest_entrypoint"
testpaths = [
    "core/tests",
    "simulator/tests",
    "experiment/tests",
]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[dependency-groups]
dev = [
    "linkify-it-py>=2.0.3",
    "myst-parser>=4.0.1",
    "sphinx>=8.2.3",
    "sphinx-rtd-theme>=3.0.2",
    "vulture>=2.14",
    "import-linter>=2.0",
    "pytest>=9.0.2",
    "deptry>=0.24.0",
]

[tool.vulture]
exclude = ["core/src/core/interfaces/dashboard.py"]
